requirements.txt
langchain
langchain-groq
python-dotenv
chromadb
langchain_community
